correct handling of empty string arguments
There is a difference between variable that's not set, and one set to an
empty string. Use -v test to avoid assumption that epmty string means
the same thing as an unset variable. This allows removing album or album
artist from scrobbles.

Get rid of all variable declarations from main() - unset variables can
now be detected and handled.

Add some logic to detect edge cases:
 - original album not set -> don't set original album artist
 - new album artist not set when changing artist ->
   assume new album artist is the same an new artist
 - new album is blank -> new album artist should be blank too
 - new album artist not set when original album not set and changing
   album -> assume new album artist is the same as original artist

This is quite complex, but user experience is better when the user
doesn't have to specify everything if an educated guess can be made.

But there's probably a ton of other edge cases regarding album artists.