hardcode last.fm API key
The application lfmedit is registered at last.fm, so I don't see a point
to have the API key configurable - better to have it hardcoded in the
only function that uses it. Don't check if the environment variable used
for storing the API key is set anymore.

Break up long line in requestOriginalScrobbleData(), and improve error
message on failure to send the API request.