split lfmedit.sh into smaller files
The monolithic lfmedit.sh is becoming hard to maintain, and makes it
difficult to write a separate "batch-mode" script.

The way the file is split is not pretty, because lfmedit.sh internally
communicates using global variables, and after splitting it's not
guaranteed that all of them will be set. New scripts must be sourced in
correct order, and care must be taken to actually provide the necessary
variables.

I think I'm approaching the limit of usefulness of bash here. The task
is becoming too complex for a shell script.