let curl perform URL encoding of requests
Remove urlEncode() function. Don't construct the request string first
and url-encode it just before sending. Instead, pass all parts of the
request to curl using multiple --data-urlencode options.

The consequence is that it's not possible to write the request in its
final "to-be-sent" form in the debug print, but since URL-encoding is
now done by curl I don't think it's a problem.