fix handling of empty fields in input
When IFS is a whitespace character (e.g. a tab, like in this case) the
empty fields are not handled correctly. Adjacent whitespace IFS chars
are treated as one separator, not as multiple separators. This does not
affect non-whitespace characters: https://stackoverflow.com/a/66480909

A workaround is to change the tab characters to non-white characters in
input, and use these as IFS. I chose Unit Separator (\037), because it
makes sense in this context.

This could be the underlying issue of #3.