introduce additional debug level
Introduce debug level 3. This sets "set -x", which produces excessive
amount of debug prints.

Change the colour of error prints to dark yellow (same as warning prints
before they were deleted). printEditData() uses red colour, and error
prints are not immediately visible because of that.

Change the location of user agent string in the source file - move it
near the top, so that it's easy to find.