make verification of edits optional
It turns out that verification can produce both false negatives and
false positives, so it is close to useless in its current form. This
can be improved if Last.fm will return album artist information in its
API results. So I'm not removing this completely, just making the
verification step disabled by default. Option -V enables it.

The order of cases in lfmedit.sh is changed to be consistent with the
list of options passed to getopt().

Debug prints are added for options -n, -Y and -V.

Closes #2.