reset album artist info in each loop iteration
Contrary to the data lfmbatchedit.sh is reading from the input file,
original album artist and new album artist were not overwritten in
each loop iteration. But they should be, otherwise once the new album
artist is set, it applies to all later edits.

New album artist was initially reset in each loop iteration, but that
was removed in 24ae068. Original album artist was not reset before.

The root cause of this problem is using global variables for holding
edit parameters instead of passing them as arguments.

Related to #3. Not sure if it fixes it, though.