avoid pausing when there are no scrobbles left to edit
When the total number of scrobbles to edit is evenly divisible by the
burst size the condition for pausing editing is triggered after editing
the last scrobble, but this is not necessary.

Add a counter of remaining scrobbles and do not pause editing when there
are no more scrobbles left to edit.