change input data format
The new format provides both old and new scrobble data in one input
record. There is no need to fetch the original scrobble data from
Last.fm anymore. This allows editing scrobbles without true timestamps.
It was not possible to fetch the original data of these scrobbles using
the API.

Guessing the album artist is still necessary. It is not very robust
here, it will fail when album artist is different than the track artist.

Add handleAlbumArtist() as a separate function. This has to be done
in two places now, so refactor.

Update the log format to match the new input format.

Bump version to 2.0.0-a1. This is a backward-incompatible change.