handle CSRF "middleware" tokens
Adapt to changes in Last.fm: obtain the "middleware" CSRF token first,
and send the obtained token in the POST request instead of the token
stored in the cookie. The "ajax=1" parameter in the POST request appears
to be mandatory now. If it is omitted the request is not rejected, but
the edit is not applied.

At first glance this "middleware" token appears to be ephemeral, but it
is not invalidated after multiple uses in POST requests and after many
hours. I suspect it is valid as long as "csrftoken" browser cookie is
valid. Therefore I'm only getting this token once per program runtime:
after getCsrfMiddlewareToken() fetches the token once, it becomes set as
a global variable and all the subsequent function calls are no-ops. This
avoids increasing the number of requests sent to Last.fm, which would
require slowing down the editing process in lfmbatchedit.sh even further
to avoid getting banned for hitting the rate limit.

Update the descriptions in auth_tokens.sample.