reduce edit delay
The edit delay set this way is only theoretical, and is further extended
by the network overhead and also by the time Last.fm backend takes to
process the request. This additional overhead is not predictable, and
its mean values ranged from 1.1 to 2.1 seconds in my testing.

With a fixed delay value the total network + backend overhead becomes
the limiting factor. In any case, I did not encounter any throttling
with the half-second fixed delay. The maximum edit rate I saw was about
2256 edits/hour (with the request overhead of 1.06s) - but that result
was an outlier.

It could be that this fixed edit delay could even be removed entirely,
but I didn't test that.